requirements.txt
Django>=1.9.6,<1.10
django-environ>=0.4.0,<0.5
django-markdown-deux>=1.0.5,<1.1
django-session-security>=2.3.2,<2.4
django_redis>=4.4.3,<4.5
djangorestframework>=3.3.3,<3.4
djangorestframework-jwt>=1.5.0,<1.6
dj-static>=0.0.6,<0.1.0
-e git+https://github.com/chop-dbhi/ehb-client.git@v1.0.2#egg=ehb_client
-e git+https://github.com/chop-dbhi/ehb-datasources.git@fb176e5f10151d98559658e7e85d465c684f8577#egg=ehb-datasources
ldap3>=1.4.0,<1.5
#psycopg2>=2.6.1,<2.7
python-json-logger==0.1.7
django-admin-tools
django-crispy-forms
